% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeCorrelationAndHeightError}
\alias{computeCorrelationAndHeightError}
\title{computeCorrelationAndHeightError}
\usage{
computeCorrelationAndHeightError(
  stemMap,
  coords = NULL,
  htField = NULL,
  plotX = 0,
  plotY = 0,
  plotRadius = 17.68,
  crs = NULL,
  treeBufferSize = 1,
  CHM = NULL,
  searchRadius = 0
)
}
\arguments{
\item{stemMap}{Data frame with stem map data for a single plot. Needs to have XY coordinates for
trees and a height for each tree.}

\item{coords}{Names of fields in \code{stemMap} containing X and Y values for tree locations.}

\item{htField}{Name of field in \code{stemMap} containing the tree height. Units should be the same
as the height units in the CHM.}

\item{plotX}{Plot easting.}

\item{plotY}{plot northing.}

\item{plotRadius}{Radius of the plot used when measuring trees.}

\item{crs}{EPSG code for the coordinate reference system.}

\item{treeBufferSize}{Size of the buffer added to each tree. Typically small but this seems to help when there
is error in relative tree locations. \code{treeBufferSize} can be a list of values so you can use a different
buffer size for each tree. This may be useful if you want to scale the area considered
when computing correlation and height error to the tree height. if \code{treeBufferSize} is 0,
single cells in the tree map are populated with the tree height (tree locations are considered as
points).}

\item{CHM}{Canopy height model in terra SpatRaster format.}

\item{searchRadius}{Maximum distance to offset plot location. This is needed to crop the CHM to the full
area that will be tested for a good location to ensure that correlation and height error metrics match
those returned by \code{testPlotLocations}.}
}
\value{
Data frame with the correlation and height error statistics. The combined error is not meaningful
 for a single location since the height error term in the combined score would go to 0.0.
}
\description{
Computes correlation between the stem map (with height) and CHM along with a
height error statistic.
}
\details{
This function is typically used in a brute force process designed to find the
best location for a plot given stem-mapped trees and a lidar-derived canopy
height model (CHM). It can also be used to compute a one-time set of fit
metrics useful for comparing one location with another.

The correlation is Pearson's correlation coefficient computed by
terra::layerCor between a tree map and the CHM (masked by the tree map raster
layer). The tree map is created using a buffer around the tree locations
(buffer can be 0) with cells for the tree populated with height values. Using
a small buffer seems to help in situations where trees lean (base XY and top
XY are different) or where the top of the tree is not best represented by a
single pixel in the CHM (possibly deciduous trees). In classic applications
that use the correlation coefficient to compare images, gray scale values are
used to compute the coefficient. In our case, we use the height values from
the CHM and tree map to compute the coefficient.

The height error is the weighted mean absolute error between the tree map and
the CHM.

The is method is taken from:
Monnet, Jean-Matthieu, and Éric Mermin. 2014. "Cross-Correlation of Diameter Measures for
the Co-Registration of Forest Inventory Plots with Airborne Laser Scanning Data" Forests 5,
no. 9: 2307-2326. https://doi.org/10.3390/f5092307

The application of Pearson's correlation coefficient is also described in:
Arthur Miranda Neto, Alessandro Corrêa Victorino, Isabelle Fantoni, Douglas Eduardo
Zampieri, Janito Vaqueiro Ferreira, et al.. Image Processing Using Pearson’s
Correlation Coefficient: Applications on Autonomous Robotics. 13th International
Conference on Mobile Robots and Competitions (Robotica 2013), Apr 2013, Lisbon,
Portugal. pp.14-19. ffhal-00860912f
}
