% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeLocalTreePositions}
\alias{computeLocalTreePositions}
\title{Compute tree positions given azimuth and distance measurements for each tree.}
\usage{
computeLocalTreePositions(
  trees,
  xRef = 0,
  yRef = 0,
  azLabel = "Azimuth",
  distLabel = "Distance",
  dbhLabel = "DBH_cm",
  dbhConversionFactor = 0.01,
  declination = 0,
  adjustForDBH = FALSE
)
}
\arguments{
\item{trees}{Data frame with tree information. Must contain at least azimuth and distance (m).
If adjusting distances to offset to the center of each tree, the data frame must contain
the DBH. By default, DBH should be in cm but you can include the \code{dbhConversionFactor}
parameter to use DBH in different units.}

\item{xRef}{X value for the reference point.}

\item{yRef}{Y value for the reference point.}

\item{azLabel}{Column label for the column containing azimuth values.}

\item{distLabel}{Column label for the column containing distance values.}

\item{dbhLabel}{Column label for the column containing DBH value.}

\item{dbhConversionFactor}{Conversion factor to convert units for tree DBH into the same units
as distances.}

\item{declination}{Magnetic declination. East declination is positive. You only need the declination
if compasses were set to 0 declination for field measurements. If compasses were set for local
declination, you don't need to provide a declination value.}

\item{adjustForDBH}{Boolean to control offsetting distance to account for tree DBH. Set this to TRUE if
distances were measured to the side of the tree facing the reference point. If disances were
measured to the center of the tree, set to FALSE.}
}
\value{
Data frame containing the tree locations. Rows are in the same order as the input \code{trees} data frame.
}
\description{
Computes tree locations given the distance and azimuth to each tree from a single reference point.
Ideally you have an accurate location for the reference point. However, you can produce a stem
map in local coordinates (reference point at (0,0)) and then shift tree locations at a later time
using the \code{moveTreesToPlotXY} function.
}
\details{
This function is designed for cases where all trees are measured from a single reference point. If
you have used multiple reference points, you can call \code{computeLocalTreePositions} multiple
times using the different \code{xRef,yRef} and associated trees and then merge the results to
create a single tree list. This function is best suited for circular plots but can be used for
any data where distance and azimuth from a known reference point have been recorded.
}
