% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computePointXY}
\alias{computePointXY}
\title{Compute location of a new point relative to a reference point}
\usage{
computePointXY(
  xRef,
  yRef,
  azimuth,
  distance,
  declination = 0,
  id = NULL,
  reverse = FALSE
)
}
\arguments{
\item{xRef}{X value for the reference point.}

\item{yRef}{Y value for the reference point.}

\item{azimuth}{azimuth in degrees from the reference point to the new point (or from the new point to the reference
point if \code{reverse = TRUE}).}

\item{distance}{distance from the reference point to the new point.}

\item{declination}{Magnetic declination. East declination is positive. You only need the declination
if compasses were set to 0 declination for field measurements. If compasses were set for local
declination, you don't need to provide a declination value.}

\item{id}{identifier for the new point. If not NULL, this will be added as the first column
in the data frame returned by \code{computePointXY}.}

\item{reverse}{boolean indicating that the azimuth is from the new point to the reference point
instead of from the reference point to the new point.}
}
\value{
data frame with the new point X and Y. If \code{id} is not NULL, \code{id} is included
 as the first column in the returned data frame.
}
\description{
\code{computePointXY} is used to compute the location of a new point given the location
of a reference point and azimuth and distance to the new point. This is useful
when a GNSS receiver is used to collect the location of a reference point but stem
locations are mapped from a different point. In the field, the crew records the
distance and azimuth from the reference point to the new point (or from the new
point to the reference point with \code{reverse = TRUE}). The location of the new
point is then used as \code{xRef, yRef} in \code{moveTreesToPlotXY} or \code{ComputeTreePositions}.
}
