% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findBestPlotLocation}
\alias{findBestPlotLocation}
\title{findBestPlotLocation}
\usage{
findBestPlotLocation(searchResults, rule = "maxcorrelation")
}
\arguments{
\item{searchResults}{Data frame returned by \code{testPlotLocations}.}

\item{rule}{Rule to be used to determine the best plot location. Possible
values are "maxCorrelation", "minHeightError", and "combined".}
}
\value{
Index into the data frame containing the offset from the original plot
location and the correlation and height error "scores".
}
\description{
findBestPlotLocation
}
