% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseCorrelation.R
\name{findBestPlotLocationPhaseCorrelation}
\alias{findBestPlotLocationPhaseCorrelation}
\title{Compute the best location for the stem map relative to the CHM using phase correlation.}
\usage{
findBestPlotLocationPhaseCorrelation(
  CHM,
  stemMap,
  initialX,
  initialY,
  searchRadius,
  stemLocationFields = c("X", "Y"),
  stemHeightField = "Ht",
  CHMbuffer = 1,
  stemMapBuffer = 1,
  cropCHM = TRUE,
  method = "buffer",
  normalize = TRUE,
  addNoise = TRUE,
  noiseMagnitude = 1
)
}
\arguments{
\item{CHM}{SpatRaster representing the canopy height surface}

\item{stemMap}{sf point object with tree positions based on \code{initialX,initialY} location}

\item{initialX}{initial reference point location}

\item{initialY}{initial reference point location}

\item{searchRadius}{radius defining area for possible locations. \code{searchRadius} will be used
to crop the \code{CHM} to limit the area for the search.}

\item{stemLocationFields}{character list with the field names in \code{stemMap} containing the X
and Y values for the tree locations.}

\item{stemHeightField}{character name of the field in \code{stemMap} containing the tree height.}

\item{CHMbuffer}{radius for each TAO detected in \code{CHM}}

\item{stemMapBuffer}{radius for each tree in the \code{stemMap}}

\item{cropCHM}{boolean to indicate whether or not the CHM should be cropped to the search area
defined by the \code{(initialX,initialY)} and \code{searchRadius}.}

\item{method}{character string defining the method used to handle the \code{CHM}. Possible
values are "buffer" and "mask".}

\item{normalize}{boolean: if TRUE, normalize the \code{CHM} and \code{stemMap} images by dividing
by the mean value of each image.}

\item{addNoise}{boolean: if TRUE, add random noise to the \code{stemMap} image. This helps
with the matching process when you have areas of high density trees with heights
similar to those on the plot.}

\item{noiseMagnitude}{numeric value form 0.0 to 1.0 that is multiplied by the minimum
and maximum values in the \code{CHM} image to set limits on the values used to add noise to
the \code{CHM} image.}
}
\value{
invisible list containing the offsets from the \code{(initialX,initialY)} position
 and the (X,Y) for the best plot location.
}
\description{
findBestPlotLocationPhaseCorrelation identifies tree approximate objects (TAOs) in the \code{CHM},
buffers the TAOs using \code{CHMbuffer}, and creates a raster image of the objects. \code{method}
controls the way the CHM is used for the image. When \code{method="buffer"}, the circular area
associated with each TAO is assigned the same height as the TAO. When \code{method="mask"}, the \code{CHM}
is masked using the TAO circles. The \code{stemMap} is used to create a second raster image representing
the trees on a plot. Each tree location is buffered using \code{stemMapBuffer} and all pixels in
the buffer are assigned the tree height.
}
\details{
The shift of the stem map image that best aligns with the CHM image is determined using phase correlation
and converted into a new coordinate for the reference point. This new coordinate can be used to translate
the trees in the \code{stemMap}.
}
