% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{isOpt}
\alias{isOpt}
\title{This is a helper function used in the package to test the
value of an object: usually a function parameter. While this simply
wraps \code{is.null()}, it makes for slightly more readable code when
checking function parameters.}
\usage{
isOpt(x)
}
\arguments{
\item{x}{object}
}
\value{
A (invisible) boolean value, TRUE indicates the object value is not NULL,
  FALSE indicates the object value is NULL.
}
\description{
This is a helper function used in the package to test the
value of an object: usually a function parameter. While this simply
wraps \code{is.null()}, it makes for slightly more readable code when
checking function parameters.
}
\keyword{internal}
