% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{moveTreesToPlotXY}
\alias{moveTreesToPlotXY}
\title{Translates trees from a local origin to a refrence point.}
\usage{
moveTreesToPlotXY(
  trees,
  xRef = 0,
  yRef = 0,
  xLabel = "Xfield",
  yLabel = "Yfield"
)
}
\arguments{
\item{trees}{Data frame with tree information. Must contain at least azimuth and distance (m).
If adjusting distances to offset to the center of each tree, the data frame must contain
the DBH. By default, DBH should be in cm but you can include the \code{dbhConversionFactor}
parameter to use DBH in different units.}

\item{xRef}{X value for the reference point.}

\item{yRef}{Y value for the reference point.}

\item{xLabel}{Column label for the column containing X values for trees.}

\item{yLabel}{Column label for the column containing Y values for trees.}
}
\value{
Return value is the \code{trees} data frame with modified X and Y values.
}
\description{
\code{moveTreesToPlotXY} assumes tree locations are relative to (0,0) and applies
a simple translation operation to shift trees. If a stem map was initially created
with an incorrect reference points, you will need to call \code{moveTreesToPlotXY}
with the incorrect X and Y multiplied by -1 to translate tree locations to a local
reference point and then call \code{moveTreesToPlotXY} again with the new reference
point location.
}
