% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rasterizeSearchResults}
\alias{rasterizeSearchResults}
\title{rasterizeSearchResults}
\usage{
rasterizeSearchResults(searchResults, value = "correlation")
}
\arguments{
\item{searchResults}{Data frame returned by \code{testPlotLocations}.}

\item{value}{Value used to populate the raster. Possible values are "correlation",
"heightError", and "combined".}
}
\value{
SpatRaster object
}
\description{
Rasterize the results of a brute force search for plot locations.
}
