% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseCorrelation.R
\name{xcorr3d}
\alias{xcorr3d}
\title{xcoor3d: perform image registration using phase correlation. This function
was taken from the imagefx package and modified to better handle NA values
in the image matrices.}
\usage{
xcorr3d(img1, img2)
}
\arguments{
\item{img1}{matrix: image representing reference set of objects}

\item{img2}{matrix: image representing movable set of objects}
}
\value{
\code{list} containing the following items:
   max.shifts shifts to align img2 with img1
   max.corr correlation values
   max.count
   corr.mat
}
\description{
xcoor3d: perform image registration using phase correlation. This function
was taken from the imagefx package and modified to better handle NA values
in the image matrices.
}
