% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{testPlotLocations}
\alias{testPlotLocations}
\title{Test a grid of potential plot locations.}
\usage{
testPlotLocations(
  stemMap,
  coords = NULL,
  htField = NULL,
  plotX = 0,
  plotY = 0,
  plotRadius = 17.68,
  crs = NULL,
  treeBufferSize = 1,
  CHM = NULL,
  searchRadius = NULL,
  progress = TRUE
)
}
\arguments{
\item{stemMap}{Data frame with stem map data for a single plot. Needs to have XY coordinates for
trees and a height for each tree.}

\item{coords}{Names of fields in stemMap containing X and Y values for tree locations.}

\item{htField}{Name of field in stemMap containing the tree height. Units should be the same
as the height units in the CHM.}

\item{plotX}{Plot easting.}

\item{plotY}{plot northing.}

\item{plotRadius}{Radius of the plot used when measuring trees.}

\item{crs}{EPSG code for the coordinate reference system.}

\item{treeBufferSize}{Size of the buffer added to each tree. Typically small but this seems to help when there
is error in relative tree locations.}

\item{CHM}{Canopy height model in terra SpatRaster format.}

\item{searchRadius}{Maximum distance to offset plot location.}

\item{progress}{Boolean indicating whether or not progress dots should be printed.}
}
\value{
Data frame with search results.
}
\description{
Apply a brute-force approach to test a grid of potential plot locations and
report the results in a data frame including the offsets from the original
plot location, correlation scores and overall height error associated with
the trial plot location. The results are used with findBestPlotLocation() to
find the best location based on the correlation or the height error results
or a combination of the two. For the combined rule, the height error is
normalized relative to the maximum height error and subtracted from 1.0. The
combined "score" is then the correlation * normalized height and the best
location is the one with the highest combined score.
}
\details{
The grid of potential plot locations is created using the resolution of the CHM so the
plotRadius should be a multiple of the CHM resolution.
}
